% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_paper.R
\name{f_sym_stage_shift}
\alias{f_sym_stage_shift}
\title{Runs the model using stated inputs}
\usage{
f_sym_stage_shift(
  symp_cluster,
  dwell,
  timepoint_of_impact,
  reduction_after,
  nsim
)
}
\arguments{
\item{symp_cluster}{# A character vector corresponding to a column in waiting_bystage_symp and sens_bystypestage}

\item{dwell}{# A character vector corresponding to a column in dwelltime_bystypestage}

\item{timepoint_of_impact}{A scalar e.g. 42, giving the number of days after which the MCT has an effect}

\item{reduction_after}{A scalar, e.g. 50, giving the percentage reduction in waiting time due to the MCT}

\item{nsim}{number of Monte Carlo simulations}
}
\value{
dataframe with stageshift percentages for each cancer type and stage combination
}
\description{
Runs the model using stated inputs
}
\examples{
f_sym_stage_shift()
Check args - these will be updated when the dwell sets and symptom groups are fully decided upon
Current thinking is one dwell set
And to limit results to three timepoint_of_impact X reduction_after combinations
An internal function which calculates the percentage stage shift for a specific cancer type and stage, flows through the above function arguments
First, get the cancer type and stage specific dwell time for the usual care stage
Next, get cancer symptom and stage specific waiting time distribution parameters (mean log and sd log)
Then, create a datrame the length of the number of simulations
and determine for each simulated patient whether they are all shifted or not
_____
This internal function is then applied to each posible cancer type and stage combination
Run for all cancer type and stage combinations
Adjust percentage stage shift for test sensitivity
tidy up
}
